Setup Screen, new game loop, almost can play
Added in the setup screen and added in most functionality for the
parameters.  Changed the game loop to allow for running in a separate
thread and allowing for human players.  Attacking functionality is
almost finished, but there may be some bugs in it.  We might want to
set a limit of attacks per turn since the map is so small